# be/cmd/swag/docs/swagger.yaml
basePath: /
definitions:
  model.AcceptFriendRequestInput:
    properties:
      request_id:
        type: string
      user_accept_id:
        type: string
    type: object
...
  model.CreateFriendRequestInput:
    properties:
      email_friend:
        type: string
      user_id:
        type: string
    type: object
...
  model.DeleteFriendInput:
    properties:
      friend_email:
        type: string
      user_id:
        type: string
    type: object
  model.EndFriendRequestInput:
    properties:
      request_id:
        type: string
      user_id:
        type: string
    type: object
...
  model.RejectFriendRequestInput:
    properties:
      request_id:
        type: string
      user_accept_id:
        type: string
    type: object
...
  model.UpdateUserInfoInput:
    properties:
      user_avatar:
        type: string
      user_birthday:
        type: string
      user_gender:
        type: string
      user_id:
        type: string
      user_nickname:
        type: string
    type: object
...
  model.UserChangePasswordInput:
    properties:
      new_password:
        type: string
      old_password:
        type: string
      user_id:
        type: string
    type: object
  model.UserFindInput:
    properties:
      limit:
        type: integer
      page:
        type: integer
      user_email:
        type: string
    type: object
...
  response.ErrResponseData:
    properties:
      code:
        description: Ma status code
        type: integer
      detail:
        description: Thong bao loi
      error:
        type: string
    type: object
  response.ResponseData:
    properties:
      code:
        description: Ma status code
        type: integer
      data:
        description: Du lieu duoc return
      message:
        description: Thong bao loi
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    email: lytranvinh.work@gmail.com
    name: TEAM V
    url: example.com/be/
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: example.com/be/
  title: API Documentation Chat App
  version: 1.0.0
paths:
...
  /api/v1/user/accept_friend_request:
    post:
      consumes:
      - application/json
      description: Accept friend request from another user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Accept friend request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AcceptFriendRequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Accept friend request
      tags:
      - User Info
  /api/v1/user/create_friend_request:
    post:
      consumes:
      - application/json
      description: Create friend request to another user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Create friend request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateFriendRequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Create friend request
      tags:
      - User Info
  /api/v1/user/delete_friend:
    delete:
      consumes:
      - application/json
      description: Delete friend user information from the service
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: delete friend user information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.DeleteFriendInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Delete friend user
      tags:
      - User Info
  /api/v1/user/end_friend_request:
    delete:
      consumes:
      - application/json
      description: End friend request to another user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: End friend request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.EndFriendRequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: End friend request
      tags:
      - User Info
  /api/v1/user/find_user:
    get:
      consumes:
      - application/json
      description: Find a user by mail address
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Find user by email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UserFindInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Find user
      tags:
      - User Info
...
  /api/v1/user/get_list_friend:
    get:
      consumes:
      - application/json
      description: Get list user friend
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Limit number
        in: query
        name: limit
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Get list user friend
      tags:
      - User Info
  /api/v1/user/get_list_friend_request:
    get:
      consumes:
      - application/json
      description: Get list friend request of user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Limit number
        in: query
        name: limit
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Get list friend request
      tags:
      - User Info
  /api/v1/user/get_user_info:
    get:
      consumes:
      - application/json
      description: Get user information after login
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Get user info
      tags:
      - User Info
...
  /api/v1/user/reject_friend_request:
    post:
      consumes:
      - application/json
      description: Reject friend request from another user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: request id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RejectFriendRequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Reject friend request
      tags:
      - User Info
  /api/v1/user/update_password:
    put:
      consumes:
      - application/json
      description: Update password user for user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update password user for user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UserChangePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Update password user for user
      tags:
      - User Info
  /api/v1/user/update_user_info:
    put:
      consumes:
      - application/json
      description: Update nick name, ... information user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update user info input date format dd-MM-yyyy HH:mm:ss
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserInfoInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Update user info
      tags:
      - User Info
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"